fix: restrict peddler endpoints

+++ apps/backend/src/app/modules/peddler/peddler.controller.ts
@@ -47,6 +47,7 @@ export class PeddlerController {
 		private readonly regionService: RegionService,
 	) {}
 
+  // public route
 	@Post()
 	@UseGuards(LoggedInGuard)
 	async create(@Body(new ValidationPipe(CreatePeddlerInputSchema)) data: CreatePeddlerInput) {
@@ -57,15 +58,23 @@ export class PeddlerController {
 	async updateById(
 		@Param("id", new ValidationPipe(NonEmptyStringSchema)) id: string,
 		@Body(new ValidationPipe(UpdatePeddlerInputSchema)) data: UpdatePeddlerInput,
+    @Roles() roles: StrictRole[],
 	) {
-		return await this.peddlerService.updateById(id, data);
+    if (rolesHavePermission(roles, "peddler", "readWrite", { id })) {
+      return await this.peddlerService.updateById(id, data);
+    }
+    throw new AppError(AppErrorTypes.NoPermission);
 	}
 
 	@Delete(":id")
-	async deleteById(@Param("id", new ValidationPipe(NonEmptyStringSchema)) id: string) {
-		return await this.peddlerService.deleteById(id);
+	async deleteById(@Param("id", new ValidationPipe(NonEmptyStringSchema)) id: string, @Roles() roles: StrictRole[]) {
+    if (rolesHavePermission(roles, "peddler", "readWrite", { id })) {
+      return await this.peddlerService.deleteById(id);
+    }
+    throw new AppError(AppErrorTypes.NoPermission);
 	}
 
+  // public route
 	@Get("all")
 	async getAll() {
 		return await this.peddlerService.getAll();
@@ -76,6 +85,7 @@ export class PeddlerController {
 		return await this.peddlerService.getById(id);
 	}
 
+  // public route
 	@Get("disability/all")
 	@UseGuards(LoggedInGuard)
 	async getAllDisabilities() {
@@ -127,6 +137,7 @@ export class PeddlerController {
 		throw new AppError(AppErrorTypes.NoPermission);
 	}
 
+  // public route
 	@Get("region/all")
 	@UseGuards(LoggedInGuard)
 	async getAllRegions() {
